+++ CHANGELOG.md
@@ -3,8 +3,15 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
-- collection subcommand
-- set collection description use case
+- `collection` sub command
+- `collection set description` sub command
+- `op set description` sub command
+- `op create` sub command
+
+### Removed
+- `add-op` sub command
+- `op set-description` sub command
+
 ### Fixed
 - engine API errors not handled